requirements.txt
Django>=4.2.20,<5.4
djangorestframework>=3.12.4,<3.16.0
djangorestframework-simplejwt>=5.0.0,<5.5.0
mysqlclient>=2.0.3,<2.2.7
drf-spectacular>=0.15.1,<0.28.0
uwsgi>=2.0.20,<2.0.28